1.1.9c Ammo Spawner Fixes
* Swapped TnH Converter menu order so ammo station is 2nd to spawn
* Rounds page hides when buying single rounds is disabled
* Rounds page only updates when open
* Fixed Ammo page showing all buy icons
* Fixed Rounds page to work
* Fixed Spawning Loot not working with Group IDs
* Removed unused Debug Log
* Added error checks to Recycler to prevent strange edge cases
* Fixed Editor Costs and timers not using float for conversion
* Fixed Compass numbers being inverted
* Fixed Compass direction numbers hiding

+++ Supply Raid Editor/Assets/Scripts/CharacterUI.cs
@@ -260,9 +260,9 @@ namespace Supply_Raid_Editor
 
             //Points auto assigned
 
-            character.newMagazineCost = int.Parse(newMagazine.text);
-            character.upgradeMagazineCost = int.Parse(upgradeMagazine.text);
-            character.duplicateMagazineCost = int.Parse(duplicateMagazine.text);
+            character.newMagazineCost = float.Parse(newMagazine.text);
+            character.upgradeMagazineCost = float.Parse(upgradeMagazine.text);
+            character.duplicateMagazineCost = float.Parse(duplicateMagazine.text);
             character.modCost = int.Parse(customMod.text);
             character.recyclerPoints = int.Parse(pointsRecycle.text);
             character.recyclerTokens = int.Parse(tokenRecycle.text);
